Rename the test-sim aux module to auxiliary.
The module directory is named aux, which is a reserved device name on Windows: libgit2, which cargo uses to check out git dependencies, refuses to write such paths, so any project pinning this repository as a git dependency cannot build or resolve on Windows.
