+++ polkadot/node/network/collator-protocol/tests/common/aux_real/prospective.rs
@@ -29,7 +29,7 @@ use polkadot_node_subsystem::{
 	overseer::Subsystem,
 };
 use polkadot_subsystem_test_sim::{
-	aux::{spawn_aux, AuxSlot},
+	auxiliary::{spawn_aux, AuxSlot},
 	harness::{Sim, SubsystemUnderTest},
 };
