+++ polkadot/node/network/collator-protocol/tests/common/mod.rs
@@ -37,11 +37,11 @@ pub mod world_helpers;
 // Re-exports preserving the previous `crate::*` paths the scenarios imported.
 pub use polkadot_subsystem_test_sim::{chain, contract, responder, runtime};
 
-/// Aux subsystems: stubs/noops from the test-sim core, plus collator-flavoured
+/// Auxiliary subsystems: stubs/noops from the test-sim core, plus collator-flavoured
 /// real-subsystem spawners (prospective + backing).
-pub mod aux {
+pub mod auxiliary {
 	pub use crate::common::aux_real::{
 		backing::CandidateBackingAux, prospective::ProspectiveParachainsAux,
 	};
-	pub use polkadot_subsystem_test_sim::aux::*;
+	pub use polkadot_subsystem_test_sim::auxiliary::*;
 }
