+++ polkadot/node/test-sim/src/lib.rs
@@ -50,7 +50,7 @@
 #![deny(missing_docs)]
 #![deny(unused_crate_dependencies)]
 
-pub mod aux;
+pub mod auxiliary;
 pub mod builders;
 pub mod chain;
 pub mod contract;
